# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NFL Survivor Pool API",
    "version": "1.0.0",
    "description": "API for NFL Survivor Pool Web App"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Entry": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "pool": {
            "type": "string"
          },
          "request": {
            "type": "string"
          },
          "entryNumber": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "eliminated"
            ]
          },
          "eliminatedWeek": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EntryWithPicks": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "pool": {
            "type": "string"
          },
          "request": {
            "type": "string"
          },
          "entryNumber": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "eliminated"
            ]
          },
          "eliminatedWeek": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "picks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Pick"
            }
          }
        }
      },
      "Pick": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "entry": {
            "type": "string"
          },
          "entryNumber": {
            "type": "integer"
          },
          "team": {
            "type": "string"
          },
          "week": {
            "type": "integer"
          },
          "game": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "away_team": {
                "type": "string"
              },
              "home_team": {
                "type": "string"
              },
              "event_date": {
                "type": "string",
                "format": "date-time"
              },
              "schedule": {
                "type": "object",
                "properties": {
                  "week": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/entries/user": {
      "get": {
        "summary": "Get all entries for the current user",
        "tags": [
          "Entries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entry"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/api/v1/entries/user/with-picks": {
      "get": {
        "summary": "Get all entries for the current user with picks",
        "tags": [
          "Entries"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "populate",
            "schema": {
              "type": "string"
            },
            "description": "Populate picks.game data"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's entries with picks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EntryWithPicks"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/api/v1/pools/{poolId}/entries": {
      "get": {
        "summary": "Get all entries for a pool",
        "tags": [
          "Entries"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entry"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Pool not found"
          }
        }
      }
    },
    "/api/v1/entries/{id}": {
      "get": {
        "summary": "Get a single entry",
        "tags": [
          "Entries"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Entry details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Entry"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Entry not found"
          }
        }
      }
    },
    "/api/v1/entries/{entryId}/picks": {
      "post": {
        "summary": "Add or update a pick for an entry",
        "tags": [
          "Entries"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "team",
                  "week"
                ],
                "properties": {
                  "team": {
                    "type": "string"
                  },
                  "week": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pick added or updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Pick"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Entry not found"
          }
        }
      }
    },
    "/api/v1/entries/{entryId}/{entryNumber}/picks/{week}": {
      "get": {
        "summary": "Get a pick for a specific week",
        "tags": [
          "Entries"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entryNumber",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "week",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pick details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Pick"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Entry or pick not found"
          }
        }
      }
    },
...
    "/api/v1/pools/{id}/entries": {
      "get": {
        "summary": "Get entries for a pool",
        "tags": [
          "Pools"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of entries for the pool",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entry"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "404": {
            "description": "Pool not found"
          }
        }
      }
    },
...
  },
...
}